src/main/java/com/tdd/katas/microservices/vehicleservice/service/VehicleServiceImpl.java
package com.tdd.katas.microservices.vehicleservice.service;

import com.tdd.katas.microservices.vehicleservice.model.CompositeVehicleData;
import com.tdd.katas.microservices.vehicleservice.repository.VehicleRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class VehicleServiceImpl implements VehicleService {

    @Autowired
    private VehicleRepository vehicleRepository;

    @Override
    public CompositeVehicleData getVehicleData(String vin) {
        return vehicleRepository.getVehicleData(vin);
    }

}
